Moved repetitive functions to shared.py
