+++ tests/unit/test_expression_tree/test_operations/test_jac_2D.py
@@ -7,11 +7,10 @@ import pybamm
 import numpy as np
 import unittest
 from scipy.sparse import eye
-from tests import get_1p1d_discretisation_for_testing
+from tests import (get_1p1d_discretisation_for_testing,
+                   multi_var_function_test,
+                   )
 
-
-def multi_var_function_test(arg1, arg2):
-    return arg1 + arg2
 
 
 class TestJacobian(TestCase):
